CRM-28: making sure we don't delete de logged user

+++ app/Livewire/Admin/Users/Delete.php
@@ -38,6 +38,12 @@ class Delete extends Component
     {
         $this->validate();
 
+        if ($this->user->is(auth()->user())) {
+            $this->addError('confirmation', "You can't delete yourself brow.");
+
+            return;
+        }
+
         $this->user->delete();
         $this->user->notify(new UserDeletedNotification());
